Gun Mechanics implemented
Player prefab - Now has a gun object that has a poiter that revolves around the player and will be used to see which direction to shoot the object in.

Gun prefab- gunn prefab - can work with drag and droppin in any scene but need an input manager to shoot.

gun controller - controls the gun
arrow controler- resposnible for rotation the arrow around the gun

LookAtMouse - wil make any object that it is assigned to look at the mouse